+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: Python Package
+name: Python Build
 
 on: [push]
 
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.5, 3.6, 3.7, 3.8]
+        python-version: [3.6, 3.7]
 
     steps:
     - uses: actions/checkout@v2
@@ -16,10 +16,18 @@ jobs:
       uses: actions/setup-python@v1
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
+    - name: Pull quilc image and qvm image
+      run: |
+        docker pull rigetti/quilc
+        docker pull rigetti/qvm
+    - name: Run qvm and quilc
+      run: |
+        docker run -itd -p 5555:5555 rigetti/quilc -R
+        docker run -itd -p 5000:5000 rigetti/qvm -S
+    - name: Install Python dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt
+        pip install -e .[development]
     - name: Lint with flake8
       run: |
         pip install flake8
