+++ mitiq/__init__.py
@@ -3,10 +3,14 @@
 This is the top level module from which functions and classes of
 Mitiq can be directly imported.
 """
+import os
 
 from .zne import Mitigator, class_mitigator, run_mitigation, fun_mitigator
 
-with open("../VERSION.txt", "r") as f:
+directory_of_this_file = os.path.dirname(os.path.abspath(__file__))
+
+
+with open(str(directory_of_this_file)+"/../VERSION.txt", "r") as f:
     __version__ = f.read().strip()
